Fix sensitivity which doesn't work without geotrek.core
