+++ geotrek/common/utils/__init__.py
@@ -152,3 +152,11 @@ def collate_c(field):
 
 def spatial_reference():
     return "{epsg_name}".format(epsg_name=SpatialReference(settings.DISPLAY_SRID).name)
+
+
+def simplify_coords(coords):
+    if isinstance(coords, (list, tuple)):
+        return [simplify_coords(coord) for coord in coords]
+    elif isinstance(coords, float):
+        return round(coords, 7)
+    raise Exception("Param is {}. Should be <list>, <tuple> or <float>".format(type(coords)))
