+++ geotrek/sensitivity/models.py
@@ -13,8 +13,7 @@ from mapentity.models import MapEntityMixin
 from mapentity.serializers import plain_text
 from geotrek.authent.models import StructureRelated
 from geotrek.common.mixins import (OptionalPictogramMixin, NoDeleteMixin, TimeStampedModelMixin, AddPropertyMixin)
-from geotrek.common.utils import intersecting, classproperty
-from geotrek.core.models import simplify_coords
+from geotrek.common.utils import intersecting, classproperty, simplify_coords
 
 
 class SportPractice(models.Model):
